More unsigned -> signed compares

+++ src/rtlib/dev_file_encod_write_wstr.c
@@ -27,7 +27,7 @@ int fb_DevFileWriteEncodWstr( FB_FILE *handle, const FB_WCHAR* buffer, size_t ch
 	if( encod_buffer != NULL )
 	{
 		/* do write */
-		if( fwrite( encod_buffer, 1, bytes, fp ) != bytes )
+		if( fwrite( encod_buffer, 1, bytes, fp ) != (size_t)bytes )
 		{
 			FB_UNLOCK();
 			return fb_ErrorSetNum( FB_RTERROR_FILEIO );
